+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "turbo-apex-debug" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.2] - 2024-10-28
+
+### Fixed
+
+- Incorrect methodName being inserted.
+
+### Added
+
+- methodName selection recognition.
+
+## [0.0.1] - 2024-08-20
 
 - Initial release